# docs/openapi.yaml
openapi: 3.0.2
info:
  title: spaCy Server
  version: '2'
  description: |
    Industrial-strength NLP via [spaCy](https://spacy.io) and [sense2vec](https://github.com/explosion/sense2vec). No
    knowledge of spaCy or sense2vec is required to use this service.

    You can use any spaCy-compatible model with this service. Different models offer different labels. In certain parts
    of the docs, you will be asked to refer to your choice of model's "label scheme". The label scheme can be found at
    https://spacy.io/models (e.g., https://spacy.io/models/en#en_core_web_sm-labels is the label scheme for the
    `en_core_web_sm` model).

    You can view the documentation for previous versions [here](https://github.com/neelkamath/spacy-server/releases).
  contact:
    name: Neel Kamath
    email: neelkamathonline@gmail.com
    url: https://github.com/neelkamath/spacy-server
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: You'll have to [run the server](https://github.com/neelkamath/spacy-server#installation) yourself.
paths:
  /ner:
    post:
      tags: [nlp]
      description: Named entity recognition. The pretrained model must have the `ner` and `parser` pipeline components
        to use this endpoint. If a sense2vec model was bundled with the service, similar phrases can also be provided.
      operationId: ner
      requestBody:
        required: true
        description: Text to process
        content:
          application/json:
            example:
              sections:
                - Net income was $9.4 million compared to the prior year of $2.7 million. Google is a big company.
                - Revenue exceeded twelve billion dollars, with a loss of $1b.
            schema:
              $ref: '#/components/schemas/NERRequest'
      responses:
        '200':
          description: Labeled text, with phrases similar to each entity
          content:
            application/json:
              example:
                data:
                  - entities:
                      - end: 1
                        end_char: 6
                        label: ORG
                        lemma: Google
                        sense2vec:
                          - phrase: ' Google'
                            similarity: 0.8996000289916992
                          - phrase: search engine
                            similarity: 0.8485999703407288
                          - phrase: DuckDuckGo
                            similarity: 0.8126999735832214
                        start: 0
                        start_char: 0
                        text: Google
                        text_with_ws: 'Google '
                      - end: 6
                        end_char: 33
                        label: PERSON
                        lemma: Sundar Pichai
                        sense2vec: []
                        start: 4
                        start_char: 20
                        text: Sundar Pichai
                        text_with_ws: Sundar Pichai
                    text: Google is headed by Sundar Pichai.
              schema:
                $ref: '#/components/schemas/NERResponse'
        '400':
          description: The pretrained model lacks the `ner` or `parser` pipeline components.
          content:
            application/json:
              examples:
                invalid_model:
                  summary: The spaCy model lacks the required pipeline components.
                  value:
                    detail: The pretrained model (en_trf_bertbaseuncased_lg) doesn't support named entity recognition.
                sense2vec_disabled:
                  summary: Similar phrases via sense2vec were requested, but a sense2vec model wasn't bundled with the
                    service.
                  value:
                    detail: There is no sense2vec model bundled with this service.
              schema:
                $ref: '#/components/schemas/InvalidModel'
  /pos:
    post:
      tags: [nlp]
      description: Part-of-speech tagging. The pretrained model needs the `parser`, `ner`, and `tagger` pipeline
        components for this endpoint to be usable.
      operationId: pos
      requestBody:
        required: true
        description: Text to process
        content:
          application/json:
            examples:
              sentences:
                summary: Supplying sentences to annotate
                value:
                  text: Google is a company. Sundar Pichai is it's CEO.
              phrase:
                summary: Supplying a phrase to annotate
                value:
                  text: Apple is looking at buying U.K. startup for $1 billion
            schema:
              $ref: '#/components/schemas/Text'
      responses:
        '200':
          description: Annotated text
          content:
            application/json:
              example:
                data:
                  - text: Google is a company.
                    tags:
                      - char_offset: 0
                        dep: nsubj
                        ent_iob: B
                        ent_type: ORG
                        head: is
                        index: 0
                        is_alpha: true
                        is_ascii: true
                        is_bracket: false
                        is_currency: false
                        is_digit: false
                        is_left_punct: false
                        is_oov: true
                        is_punct: false
                        is_quote: false
                        is_right_punct: false
                        is_space: false
                        is_stop: false
                        is_title: true
                        lang: en
                        left_edge: Google
                        lemma: Google
                        like_email: false
                        like_num: false
                        like_url: false
                        normalized: google
                        pos: PROPN
                        prefix: G
                        prob: -20.0
                        right_edge: Google
                        shape: Xxxxx
                        suffix: gle
                        tag: NNP
                        text: Google
                        text_with_ws: 'Google '
                        whitespace: ' '
              schema:
                $ref: '#/components/schemas/PartsOfSpeech'
        '400':
          description: The pretrained model lacks the `ner`, `parser`, or `tagger` pipeline components.
          content:
            application/json:
              example:
                detail: The pretrained model (en_trf_bertbaseuncased_lg) doesn't support part-of-speech tagging.
              schema:
                $ref: '#/components/schemas/InvalidModel'
  /tokenizer:
    post:
      tags: [nlp]
      description: Tokenization
      operationId: tokenizer
      requestBody:
        required: true
        description: Text to tokenize
        content:
          application/json:
            example:
              text: Apple is looking at buying U.K. startup for $1 billion
            schema:
              $ref: '#/components/schemas/Text'
      responses:
        '200':
          description: Tokenized text
          content:
            application/json:
              example:
                tokens: [Apple, is, looking, at, buying, U.K., startup, for, $, '1', billion]
              schema:
                $ref: '#/components/schemas/Tokens'
  /sentencizer:
    post:
      tags: [nlp]
      description: Sentence segmentation. The pretrained model must have the `parser` pipeline component for this
        endpoint to be available.
      operationId: sentencizer
      requestBody:
        required: true
        description: Sentences to segmentize
        content:
          application/json:
            example:
              text: Apple is looking at buying U.K. startup for $1 billion. Another sentence.
            schema:
              $ref: '#/components/schemas/Text'
      responses:
        '200':
          description: Sentences
          content:
            application/json:
              example:
                sentences: [Apple is looking at buying U.K. startup for $1 billion., Another sentence.]
              schema:
                $ref: '#/components/schemas/Sentences'
        '400':
          description: The pretrained model lacks the `parser` pipeline component.
          content:
            application/json:
              example:
                detail: The pretrained model (en_trf_bertbaseuncased_lg) doesn't support sentence segmentation.
              schema:
                $ref: '#/components/schemas/InvalidModel'
  /health_check:
    get:
      tags: [status]
      description: Check if all systems are operational
      operationId: healthCheck
      responses:
        '204':
          description: All systems are operational
components:
  schemas:
    NERRequest:
      type: object
      properties:
        sections:
          description:
            Although you could pass the full text as a single array item, it would be faster to split large text into
            multiple items. Each item needn't be semantically related.
          type: array
          items:
            type: string
        sense2vec:
          description: Whether to also compute similar phrases using sense2vec (significantly slower)
          type: boolean
          default: false
      required: [sections]
    NERResponse:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
            properties:
              entities:
                type: array
                items:
                  type: object
                  properties:
                    text:
                      type: string
                      description: A unicode representation of the entity text.
                    label:
                      description: &nerDesc |
                        The named entity label. The model's label scheme lists which are supported. An explanation of
                        the different labels can be found [here](https://spacy.io/api/annotation#named-entities).
                      type: string
                    start_char:
                      type: integer
                      description: The character offset for the start of the entity.
                    end_char:
                      type: integer
                      description: The character offset for the end of the entity.
                    lemma:
                      type: string
                      description: The entity’s lemma.
                    sense2vec:
                      type: array
                      description: Phrases similar to the entity (empty if sense2vec was disabled)
                      items:
                        type: object
                        properties:
                          phrase:
                            type: string
                          similarity:
                            type: number
                            description: Similarity in the range of 0-1
                        required: [phrase, similarity]
                    start:
                      type: integer
                      description: The token offset for the start of the entity.
                    end:
                      type: integer
                      description: The token offset for the end of the entity.
                    text_with_ws:
                      type: string
                      description:
                        The text content of the entity with a trailing whitespace character if the last token
                        has one.
                  required: [start, end, text, text_with_ws, start_char, end_char, lemma, sense2vec, label]
              text:
                type: string
            required: [entities, text]
      required: [data]
    Text:
      type: object
      properties:
        text:
          type: string
      required: [text]
    PartsOfSpeech:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
            properties:
              text:
                type: string
                description: The sentence or phrase being tagged.
              tags:
                type: array
                items:
                  type: object
                  properties:
                    char_offset:
                      type: integer
                      description: The character offset of the token within the parent document.
                    dep:
                      type: string
                      description: The dependency label. The model's label scheme lists which are supported. An
                        explanation of the different labels can be found
                        [here](https://spacy.io/api/annotation#dependency-parsing).
                    ent_iob:
                      type: string
                      enum: [B, I, O, '']
                      description:
                        IOB code of named entity tag. `“B”` means the token begins an entity, `“I”` means it is
                        inside an entity, `“O”` means it is outside an entity, and `""` means no entity tag is set.
                    ent_type:
                      type: string
                      description: *nerDesc
                    head:
                      type: string
                      description: The syntactic parent, or “governor”, of this token.
                    index:
                      type: integer
                      description: The index of the token within the parent document.
                    is_alpha:
                      type: boolean
                      description: Does the token consist of alphabetic characters?
                    is_ascii:
                      type: boolean
                      description: Does the token consist of ASCII characters?
                    is_bracket:
                      type: boolean
                      description: Is the token a bracket?
                    is_currency:
                      type: boolean
                      description: Is the token a currency symbol?
                    is_digit:
                      type: boolean
                      description: Does the token consist of digits?
                    is_left_punct:
                      type: boolean
                      description: Is the token a left punctuation mark (e.g. `(`)?
                    is_oov:
                      type: boolean
                      description: Is the token out-of-vocabulary?
                    is_punct:
                      type: boolean
                      description: Is the token punctuation?
                    is_quote:
                      type: boolean
                      description: Is the token a quotation mark?
                    is_right_punct:
                      type: boolean
                      description: Is the token a right punctuation mark (e.g. `)`)?
                    is_space:
                      type: boolean
                      description: Does the token consist of whitespace characters?
                    is_stop:
                      type: boolean
                      description: Is the token part of a “stop list”?
                    is_title:
                      type: boolean
                      description: Is the token in titlecase?
                    lang:
                      type: string
                      enum:
                        - af
                        - ar
                        - bg
                        - bn
                        - ca
                        - cs
                        - da
                        - de
                        - el
                        - en
                        - es
                        - et
                        - fa
                        - fi
                        - fr
                        - ga
                        - he
                        - hi
                        - hr
                        - hu
                        - id
                        - is
                        - it
                        - ja
                        - kn
                        - ko
                        - lt
                        - lv
                        - mr
                        - nb
                        - nl
                        - pl
                        - pt
                        - ro
                        - ru
                        - si
                        - sk
                        - sl
                        - sq
                        - sr
                        - sv
                        - ta
                        - te
                        - th
                        - tl
                        - tr
                        - tt
                        - uk
                        - ur
                        - vi
                        - xx
                        - zh
                      description:
                        The [ISO 639-1 code](https://en.wikipedia.org/wiki/List_of_ISO_639-1_codes) for the
                        language of the parent document’s vocabulary.
                    left_edge:
                      type: string
                      description: The leftmost token of this token’s syntactic descendants.
                    lemma:
                      type: string
                      description: Base form of the token, with no inflectional suffixes.
                    like_email:
                      type: boolean
                      description: Does the token resemble an email address?
                    like_num:
                      type: boolean
                      description: Does the token represent a number (e.g. `10.9`, `10`, `"ten"`)?
                    like_url:
                      type: boolean
                      description: Does the token resemble a URL?
                    normalized:
                      type: string
                      description: The token’s norm (i.e., a normalized form of the token text).
                    pos:
                      type: string
                      description: Part-of-speech tags for the model
                    prefix:
                      type: string
                      description: A length-N substring from the start of the token. Defaults to `N=1`.
                    prob:
                      type: number
                      description:
                        Smoothed log probability estimate of token’s word type (context-independent entry in
                        the vocabulary).
                    right_edge:
                      type: string
                      description: The rightmost token of this token’s syntactic descendants.
                    shape:
                      type: string
                      description: Transform of the tokens’s string, to show orthographic features (e.g., `“Xxxx”` or
                        `“dd”`).
                    suffix:
                      type: string
                      description: Length-N substring from the end of the token. Defaults to `N=3`.
                    tag:
                      type: string
                      description: Part-of-speech tags from the model's label scheme
                    text:
                      type: string
                      description: Verbatim text content.
                    text_with_ws:
                      type: string
                      description: Text content, with trailing space character if present.
                    whitespace:
                      type: string
                      enum:  [' ', '']
                      description: Trailing space character if present.
                  required:
                    - text
                    - text_with_ws
                    - whitespace
                    - head
                    - left_edge
                    - right_edge
                    - index
                    - ent_type
                    - ent_iob
                    - lemma
                    - normalized
                    - shape
                    - prefix
                    - suffix
                    - is_alpha
                    - is_ascii
                    - is_digit
                    - is_title
                    - is_punct
                    - is_left_punct
                    - is_right_punct
                    - is_space
                    - is_bracket
                    - is_quote
                    - is_currency
                    - like_url
                    - like_num
                    - like_email
                    - is_oov
                    - is_stop
                    - pos
                    - tag
                    - dep
                    - lang
                    - prob
                    - char_offset
            required: [text, tags]
      required: [data]
    Tokens:
      type: object
      properties:
        tokens:
          type: array
          items:
            type: string
      required: [tokens]
    Sentences:
      type: object
      properties:
        sentences:
          type: array
          items:
            type: string
      required: [sentences]
    InvalidModel:
      type: object
      properties:
        detail:
          type: string
      required: [detail]